refactor stripe style builder
