+++ shared/utils/color.js
@@ -1,0 +1,10 @@
+export function rgbToHex(rgb) {
+  if (typeof rgb !== 'string') return rgb;
+  if (rgb.startsWith('#')) return rgb;
+  const match = rgb.match(/\d+/g);
+  if (!match || match.length < 3) return rgb;
+  const [r, g, b] = match.map(Number);
+  return `#${r.toString(16).padStart(2, '0')}${g
+    .toString(16)
+    .padStart(2, '0')}${b.toString(16).padStart(2, '0')}`;
+}
